Fixed various errors (wrong casts, missing includes, glib changes); compiles again now.
git-svn-id: https://fragistics.svn.sourceforge.net/svnroot/fragistics/trunk@51 2a05a446-2b2c-4a10-9532-17218a4ad73e

+++ src/crosstable.cpp
@@ -357,9 +357,9 @@ int
 __cdecl
 #endif
 intcomp(const void* p1,const void* p2){
-	int num1,num2;
-	num1=(int)p1;
-	num2=(int)p2;
+	long num1,num2;
+	num1=(long)p1;
+	num2=(long)p2;
 	if(p1<p2){
 		return 1;
 	}else{
